requirements.txt
django
djangorestframework
drf-extensions
drf_yasg
gunicorn
whitenoise